[FLINK-18957] Implement PhysicalSlotProvider and PhysicalSlotRequestBulkChecker in SlotProvider and SchedulerImpl
At the moment, only SlotProvider (SchedulerImpl) is available to create ExecutionSlotAllocatorFactory/ExecutionSlotAllocator.
SlotSharingExecutionSlotAllocator requires PhysicalSlotProvider and PhysicalSlotRequestBulkChecker.
As we need SlotPool for PhysicalSlotProvider and PhysicalSlotRequestBulkChecker, we now implement them in SlotProvider/SchedulerImpl.
This is not optimal in a long term but as SlotProvider/SchedulerImpl should become obsolete eventually,
we can later refactor the creation of components.